Fix for FNF chords showing as split notes in the path notation.
Currently CHOpt has FNF notes not counting chords properly, this is because it considers chords split notes since the actual game allows you to hit the individual notes in chords separately on non-pro instruments.

This leads to the path notation to be really hard to follow across all instruments. Removing the split note check for Festival fixes this issue.